requirements-dev.txt
# Include production dependencies
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
httpx>=0.25.0

# Code quality
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.7.0

# Type stubs
pandas-stubs>=2.0.0
types-openpyxl>=3.1.0

# Security
safety>=2.3.0
bandit>=1.7.5

# Development tools
ipython>=8.17.0
jupyter>=1.0.0

# Pre-commit hooks
pre-commit>=3.5.0
